# src/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: RentalFlow API
  description: API documentation for the RentalFlow microservices ecosystem.
  version: 1.0.0
servers:
  - url: https://rentalflow.onrender.com
    description: Production Server (API Gateway)
  - url: http://localhost:8080
    description: Local Development Server (API Gateway)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/items:
    get:
      summary: List or get items
      tags: [Inventory]
      parameters:
        - name: id
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Success
    post:
      summary: Create an item
      tags: [Inventory]
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Created
...
  /api/notifications/user:
    get:
      summary: Get user notifications
      tags: [Notification]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: user_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Success